refactor: переименована папка страницы регистрации, функции переписаны через коллбэки, улучшен стиль кода
